requirements.txt
apache-airflow
sqlparse
apache-airflow-providers-postgres 
pandas